+++ src/simulation.cpp
@@ -51,7 +51,7 @@ GCMCSimulation::~GCMCSimulation() {
     }
 }
 
-void GCMCSimulation::simulate(int steps, int start_step) {
+void GCMCSimulation::simulate(long int steps, int start_step) {
 
     for (int step {start_step + 1}; step != (steps + start_step + 1); step ++) {
         unique_ptr<MCMovetype> movetype {select_movetype()};
